INC-612: Extract out caching class into separate file without changes
