+++ server/routes/analyticsRouter.test.ts
@@ -4,7 +4,7 @@ import request from 'supertest'
 import PrisonRegister from '../data/prisonRegister'
 import config from '../config'
 import ZendeskClient from '../data/zendeskClient'
-import { StitchedTablesCache } from '../services/analyticsService'
+import StitchedTablesCache from '../services/stitchedTablesCache'
 import type { ChartId } from './analyticsChartTypes'
 import { protectedCharacteristicRoutes } from './analyticsRouter'
 import { appWithAllRoutes } from './testutils/appSetup'
